Close #12; small corrections done

+++ app/src/main/java/com/wcsm/shopperrotas/data/dto/RideEstimateResponse.kt
@@ -3,7 +3,7 @@ package com.wcsm.shopperrotas.data.dto
 data class RideEstimateResponse(
     val origin: Location,
     val destination: Location,
-    val distance: Double,
+    val distance: Int,
     val duration: String,
     val options: List<RideOption>,
     val routeResponse: Any
